Organized movie search.
Removed search functionality from MovieListView.
Created MovieSearchView to contain search logic.
Initialized new MovieSearchView in app.js to listen to search button.
