+++ src/app.js
@@ -4,7 +4,9 @@
 import $ from 'jquery';
 import _ from 'underscore';
 import MovieList from './collections/movie_list';
+import SearchList from './collections/search_list';
 import MovieListView from './views/movie_list_view';
+import MovieSearchView from './views/movie_search_view';
 
 
 var movieListTemplate = _.template($('#movie-list-template').html());
@@ -14,6 +16,12 @@ var movieList = new MovieList();
 
 var movieListView = new MovieListView({
   model: movieList,
+  movieListTemplate: movieListTemplate,
+  movieTemplate: movieTemplate,
+  el: $("#application")
+});
+
+var movieSearchView = new MovieSearchView({
   movieListTemplate: movieListTemplate,
   movieTemplate: movieTemplate,
   el: $("#application")
